requirements.txt
streamlit
pdfplumber
sentence-transformers
scikit-learn
plotly
numpy
python-docx